refactor DeleteGroupPopup and DeleteCoursePopup
