+++ src/test/resources/features/editGroup.feature
@@ -1,7 +1,8 @@
 Feature: Groups
 
+  @deleteGroupInstructor1
   Scenario: Edit a group
-    Given I log in as "Instructor02" user
+    Given I log in as "Instructor01" user
     And I create a group with:
       | name        | SeleniumWebDriver |
       | description | Description       |
